+++ shuup/admin/static_src/wizard/js/wizard.js
@@ -43,11 +43,11 @@
                 '<div class="action-bar">' +
                     '<div class="clearfix">' +
                         getButton(gettext("Previous"), "previous", disablePrevious, "btn-primary pull-left " + (config.hidePrevious? "hidden":"")) +
-                        getButton(gettext("Next"), "next", disableNext, "btn-primary pull-right") +
+                        getButton(isLastPane() ? gettext("Finish") : gettext("Next"), "next", disableNext, "btn-primary pull-right") +
                     '</div>' +
-                    (config.skip?
+                    ($activeWizardPane.data("can_skip") === "True" || config.skip?
                     '<div class="clearfix">' +
-                        getButton(gettext("Skip Setup"), "skip", false, "btn-default pull-right", config.skipTooltip) +
+                        getButton(gettext("Skip"), "skip", false, "btn-default pull-right", config.skipTooltip) +
                     '</div>': '') +
                 '</div>'
             );
@@ -87,6 +87,7 @@
             switchToPane(config);
         } else {
             this.on("click", "button[name='next']", () => {
+                this.find("button[name='next']").prepend('<i class="fa fa-spinner fa-pulse fa-fw"></i>');
                 if(config.next) {
                     config.next($activeWizardPane);
                 } else {
@@ -96,11 +97,15 @@
                         } else {
                             next();
                         }
-                    }).fail((err) => renderFormErrors($activeWizardPane.find("form"), err.responseJSON));
+                    }).fail((err) => {
+                        this.find("i.fa.fa-spinner").remove();
+                        renderFormErrors($activeWizardPane.find("form"), err.responseJSON);
+                    });
                 }
             });
 
             this.on("click", "button[name='previous']", () => {
+                this.find("button[name='previous']").prepend('<i class="fa fa-spinner fa-pulse fa-fw"></i>');
                 if(config.previous) {
                     config.previous($activeWizardPane);
                 } else {
@@ -109,7 +114,11 @@
             });
 
             this.on("click", "button[name='skip']", () => {
-                config.skip();
+                if(config.redirectOnLastPane && isLastPane()){
+                    window.location = config.redirectOnLastPane;
+                } else {
+                    next();
+                }
             });
             switchToPane(0);
         }