+++ shuup_tests/admin/test_product_module.py
@@ -13,6 +13,7 @@ from shuup.admin.modules.products.views.edit import ProductEditView
 from shuup.admin.utils.urls import get_model_url
 from shuup.admin.views.search import get_search_results
 from shuup.core.models import ProductVisibility
+from shuup.importer.admin_module import ImportAdminModule
 from shuup.testing.factories import (
     create_product, get_default_product, get_default_shop
 )
@@ -26,7 +27,7 @@ def test_product_module_search(rf, admin_user):
     get_default_shop()
     request = apply_request_middleware(rf.get("/"), user=admin_user)
 
-    with replace_modules([ProductModule]):
+    with replace_modules([ImportAdminModule, ProductModule]):
         with admin_only_urls():
             default_product = get_default_product()
             model_url = get_model_url(default_product)
@@ -45,7 +46,7 @@ def test_product_edit_view_works_at_all(rf, admin_user):
     shop_product.save()
     request = apply_request_middleware(rf.get("/"), user=admin_user)
 
-    with replace_modules([ProductModule]):
+    with replace_modules([ImportAdminModule, ProductModule]):
         with admin_only_urls():
             view_func = ProductEditView.as_view()
             response = view_func(request, pk=product.pk)
@@ -62,7 +63,7 @@ def test_product_edit_view_with_params(rf, admin_user):
     name = "test name"
     request = apply_request_middleware(rf.get("/", {"name": name, "sku": sku}), user=admin_user)
 
-    with replace_modules([ProductModule]):
+    with replace_modules([ImportAdminModule, ProductModule]):
         with admin_only_urls():
             view_func = ProductEditView.as_view()
             response = view_func(request)