+++ doc/changelog.rst
@@ -21,6 +21,10 @@ Localization
 Admin
 ~~~~~
 
+- Add option to skip wizard panes
+- Add option to return home view
+- List wizard phases at home view
+
 Addons
 ~~~~~~